requirements.txt
acres==0.5.0
aiofiles==24.1.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.3.0
attrdict==2.0.1
attrs==26.1.0
babel==2.18.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==1.3.0
cachetools==7.0.1
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.7
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.7
cssselect==1.4.0
cssutils==2.11.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
Cython==3.2.4
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
einops==0.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.59.2
fqdn==1.5.1
fsspec==2025.9.0
ftfy==6.3.1
fugashi==1.5.2
future==1.0.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.10
ImageIO==2.37.2
imagesize==2.0.0
imgaug==0.4.0
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.3.1
lazy_loader==0.4
lmdb==1.7.5
looseversion==1.3.0
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
modelscope==1.35.3
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.13
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas==13.0.2.14
nvidia-cuda-runtime==13.0.88
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.61
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-nvjitlink==13.1.80
openai==2.30.0
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
paddleocr==3.4.0
paddlepaddle==2.6.2
paddlepaddle-gpu==3.3.1
paddlex==3.4.3
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.6
pathlib==1.0.1
pdf2docx==0.5.8
pdfminer.six==20250506
pillow==12.2.0
plac==1.4.5
platformdirs==4.9.2
portalocker==3.2.0
premailer==3.10.0
preshed==3.0.10
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==3.20.2
prov==2.1.1
psutil==7.2.2
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.20.2
pyparsing==3.2.3
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==4.1.0
pytz==2025.2
pywin32==311
pywinpty==3.0.3
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.3
rarfile==4.2
rdflib==7.1.4
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
sacrebleu==2.6.0
sacremoses==0.1.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==2.1.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8.3
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tabulate==0.10.0
tensorrt_cu13==10.16.1.11
tensorrt_cu13_bindings==10.16.1.11
tensorrt_cu13_libs==10.16.1.11
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0+cu130
torchvision==0.26.0+cu130
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==5.5.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.12.0
unidic-lite==1.0.8
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.35.0
visualdl==2.5.3
wasabi==0.10.1
wcwidth==0.6.0
weasel==0.4.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
wrapt==1.17.3
